Move DecommissionAttributeMetadata to decommission package
