+++ server/src/test/java/org/opensearch/cluster/coordination/JoinTaskExecutorTests.java
@@ -37,10 +37,9 @@ import org.opensearch.cluster.ClusterName;
 import org.opensearch.cluster.ClusterState;
 import org.opensearch.cluster.ClusterStateTaskExecutor;
 import org.opensearch.cluster.decommission.DecommissionAttribute;
-import org.opensearch.cluster.decommission.DecommissionFailedException;
+import org.opensearch.cluster.decommission.DecommissionAttributeMetadata;
 import org.opensearch.cluster.decommission.DecommissionStatus;
 import org.opensearch.cluster.decommission.NodeDecommissionedException;
-import org.opensearch.cluster.metadata.DecommissionAttributeMetadata;
 import org.opensearch.cluster.metadata.IndexMetadata;
 import org.opensearch.cluster.metadata.Metadata;
 import org.opensearch.cluster.node.DiscoveryNode;
@@ -58,7 +57,6 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.Map;
 
-import static java.util.Collections.emptyMap;
 import static org.hamcrest.Matchers.is;
 import static org.opensearch.test.VersionUtils.allVersions;
 import static org.opensearch.test.VersionUtils.maxCompatibleVersion;
@@ -231,7 +229,7 @@ public class JoinTaskExecutorTests extends OpenSearchTestCase {
         Metadata.Builder metaBuilder = Metadata.builder();
         Metadata metadata = metaBuilder.build();
         DiscoveryNode discoveryNode = newDiscoveryNode(Collections.singletonMap("zone", "zone-2"));
-        JoinTaskExecutor.ensureNodeNotDecommissioned(discoveryNode, metadata);
+        JoinTaskExecutor.ensureNodeCommissioned(discoveryNode, metadata);
     }
 
     public void testPreventJoinClusterWithDecommission() {
@@ -252,7 +250,7 @@ public class JoinTaskExecutorTests extends OpenSearchTestCase {
         DiscoveryNode discoveryNode = newDiscoveryNode(Collections.singletonMap("zone", "zone-1"));
         expectThrows(
             NodeDecommissionedException.class,
-            () -> JoinTaskExecutor.ensureNodeNotDecommissioned(discoveryNode, metadata)
+            () -> JoinTaskExecutor.ensureNodeCommissioned(discoveryNode, metadata)
         );
     }
 
@@ -269,7 +267,7 @@ public class JoinTaskExecutorTests extends OpenSearchTestCase {
         Metadata metadata = metaBuilder.build();
 
         DiscoveryNode discoveryNode = newDiscoveryNode(Collections.singletonMap("zone", "zone-2"));
-        JoinTaskExecutor.ensureNodeNotDecommissioned(discoveryNode, metadata);
+        JoinTaskExecutor.ensureNodeCommissioned(discoveryNode, metadata);
     }
 
     public void testJoinClusterWithDecommissionFailedOrInitOrRecommission() {
@@ -289,7 +287,7 @@ public class JoinTaskExecutorTests extends OpenSearchTestCase {
         Metadata metadata = metaBuilder.build();
 
         DiscoveryNode discoveryNode = newDiscoveryNode(Collections.singletonMap("zone", "zone-1"));
-        JoinTaskExecutor.ensureNodeNotDecommissioned(discoveryNode, metadata);
+        JoinTaskExecutor.ensureNodeCommissioned(discoveryNode, metadata);
     }
 
     private DiscoveryNode newDiscoveryNode(Map<String, String> attributes) {
