+++ server/src/test/java/org/opensearch/cluster/metadata/DecommissionAttributeMetadataSerializationTests.java
@@ -11,6 +11,7 @@ package org.opensearch.cluster.metadata;
 import org.opensearch.cluster.ClusterModule;
 import org.opensearch.cluster.Diff;
 import org.opensearch.cluster.decommission.DecommissionAttribute;
+import org.opensearch.cluster.decommission.DecommissionAttributeMetadata;
 import org.opensearch.cluster.decommission.DecommissionStatus;
 import org.opensearch.common.io.stream.NamedWriteableRegistry;
 import org.opensearch.common.io.stream.Writeable;
